requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
h5py>=3.9.0
tqdm>=4.65.0
aiofiles>=23.1.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
streamlit>=1.25.0
plotly>=5.15.0
folium>=0.14.0
geopy>=2.3.0
exifread>=3.0.0